Data.GenericLens.Internal

Description

The library internals are exposed through this module. Please keep in mind that everything here is subject to change irrespective of the the version numbers.

Documentation

module Data.Generics.Internal.Families

module Data.Generics.Internal.Families.Changing

module Data.Generics.Internal.Families.Collect

module Data.Generics.Internal.Families.Has

module Data.Generics.Internal.Void

module Data.Generics.Internal.Errors

data family Param :: Nat -> j -> k

newtype Rec (p :: Type) a x

Constructors

Rec

Fields

unRec :: K1 R a x

class (Coercible (Rep a) (RepN a), Generic a) => GenericN (a :: Type) where

Associated Types

type RepN (a :: Type) :: Type -> Type

type RepN a = Rep (Indexed a 0)

Methods

toN :: RepN a x -> a

fromN :: a -> RepN a x

Profunctor optics

type Iso s t a b = forall p i. Profunctor p => p i a b -> p i s t

type Iso' s a = Iso s s a a

repIso :: (Generic a, Generic b) => Iso a b (Rep a x) (Rep b x)

A type and its generic representation are isomorphic

mIso :: Iso (M1 i c f p) (M1 i c g p) (f p) (g p)

M1 is just a wrapper around `f p` mIso :: Iso' (M1 i c f p) (f p)

kIso :: Iso (K1 r a p) (K1 r b p) a b

recIso :: Iso (Rec r a p) (Rec r b p) a b

sumIso :: Iso ((a :+: b) x) ((a' :+: b') x) (Either (a x) (b x)) (Either (a' x) (b' x))

prodIso :: Iso ((a :*: b) x) ((a' :*: b') x) (a x, b x) (a' x, b' x)

assoc3 :: Iso ((a, b), c) ((a', b'), c') (a, (b, c)) (a', (b', c'))

fromIso :: Iso s t a b -> Iso b a t s

iso :: (s -> a) -> (b -> t) -> Iso s t a b

withIso :: Iso s t a b -> ((s -> a) -> (b -> t) -> r) -> r

pairing :: Iso s t a b -> Iso s' t' a' b' -> Iso (s, s') (t, t') (a, a') (b, b')

type Lens s t a b = forall p i. Strong p => p i a b -> p i s t

type LensLike p s t a b = p a b -> p s t

ravel :: (ALens a b i a b -> ALens a b i s t) -> Lens s t a b

set :: ((a -> b) -> s -> t) -> (s, b) -> t

Setting

view :: Lens s s a a -> s -> a

withLensPrim :: Lens s t a b -> (forall c. (s -> (c, a)) -> ((c, b) -> t) -> r) -> r

idLens :: ALens a b i a b

first :: Lens ((a :*: b) x) ((a' :*: b) x) (a x) (a' x)

Lens focusing on the first element of a product

second :: Lens ((a :*: b) x) ((a :*: b') x) (b x) (b' x)

Lens focusing on the second element of a product

fork :: (a -> b) -> (a -> c) -> a -> (b, c)

data Coyoneda f b

Constructors

forall a. Coyoneda (a -> b) (f a)

inj :: Functor f => Coyoneda f a -> f a

proj :: Functor f => f a -> Coyoneda f a

(??) :: Functor f => f (a -> b) -> a -> f b

assoc3L :: Lens ((a, b), c) ((a', b'), c') (a, (b, c)) (a', (b', c'))

stron :: (Either s s', b) -> Either (s, b) (s', b)

choosing :: forall s t a b s' t'. Lens s t a b -> Lens s' t' a b -> Lens (Either s s') (Either t t') a b

lens :: (s -> (c, a)) -> ((c, b) -> t) -> Lens s t a b

data ALens a b i s t

Constructors

forall c. ALens (s -> (c, a)) ((c, b) -> t)

swap :: (a, b) -> (b, a)

type APrism i s t a b = Market a b i a b -> Market a b i s t

type Prism s t a b = forall p i. Choice p => p i a b -> p i s t

type Prism' s a = forall p i. Choice p => p i a a -> p i s s

left :: Prism ((a :+: c) x) ((b :+: c) x) (a x) (b x)

right :: Prism ((a :+: b) x) ((a :+: c) x) (b x) (c x)

prism :: (b -> t) -> (s -> Either t a) -> Prism s t a b

_Left :: Prism (Either a c) (Either b c) a b

_Right :: Prism (Either c a) (Either c b) a b

prismPRavel :: APrism i s t a b -> Prism s t a b

build :: (Tagged i b b -> Tagged i t t) -> b -> t

match :: Prism s t a b -> s -> Either t a

without' :: Prism s t a b -> Prism s t c d -> Prism s t (Either a c) (Either b d)

withPrism :: APrism i s t a b -> ((b -> t) -> (s -> Either t a) -> r) -> r

prism2prismp :: Market a b i s t -> Prism s t a b

idPrism :: Market a b i a b

gsum :: (a x -> c) -> (b x -> c) -> (a :+: b) x -> c

module Data.Generics.Product.Internal.Subtype