Data.Generics.Product.Internal.Fields

Documentation

type Context_ field s t a b = (HasTotalFieldP field (Rep s) ~ 'Just a, HasTotalFieldP field (Rep t) ~ 'Just b, UnifyHead s t, UnifyHead t s)

type Context' field s a = (Generic s, ErrorUnless field s (CollectField field (Rep s)), GLens' (HasTotalFieldPSym field) (Rep s) a, Defined (Rep s) (NoGeneric s '[(('Text "arising from a generic lens focusing on the " ':<>: QuoteType field) ':<>: 'Text " field of type ") ':<>: QuoteType a, 'Text "in " ':<>: QuoteType s]) (() :: Constraint))

type Context0 field s t a b = (Generic s, Generic t, GLens (HasTotalFieldPSym field) (Rep s) (Rep t) a b, ErrorUnless field s (CollectField field (Rep s)), Defined (Rep s) (NoGeneric s '[(('Text "arising from a generic lens focusing on the " ':<>: QuoteType field) ':<>: 'Text " field of type ") ':<>: QuoteType a, 'Text "in " ':<>: QuoteType s]) (() :: Constraint))

class Context (field :: Symbol) s t a b | s field -> a, t field -> b, s field b -> t, t field a -> s

derived :: forall field s t a b. Context0 field s t a b => Lens s t a b