Data.Generics.Product.Internal.HList

Description

Derive an isomorphism between a product type and a flat HList.

Documentation

class GIsList (f :: Type -> Type) (g :: Type -> Type) (as :: [Type]) (bs :: [Type]) | f -> as, g -> bs, bs f -> g, as g -> f where

Methods

glist :: Iso (f x) (g x) (HList as) (HList bs)

class IndexList (i :: Nat) as bs a b | i as -> a, i bs -> b, i as b -> bs, i bs a -> as where

Methods

point :: Lens (HList as) (HList bs) a b

data HList (as :: [Type]) where

Constructors

Nil :: HList '[]

(:>) :: a -> HList as -> HList (a ': as) infixr 5

type family (as :: [k]) ++ (bs :: [k]) :: [k] where ...

Equations

'[] ++ bs = bs

(a ': as) ++ bs = a ': (as ++ bs)

class Elem (as :: [(k, Type)]) (key :: k) (i :: Nat) a | as key -> i a

class ListTuple (tuple :: Type) (tuple' :: Type) (as :: [Type]) (bs :: [Type]) | as -> tuple, bs -> tuple' where

Minimal complete definition

tupleToList, listToTuple

Methods

tupled :: Iso (HList as) (HList bs) tuple tuple'

tupleToList :: tuple' -> HList bs

listToTuple :: HList as -> tuple

type family TupleToList a where ...

Equations

TupleToList () = '[]

TupleToList (a, b) = '[a, b]

TupleToList (a, b, c) = '[a, b, c]

TupleToList (a, b, c, d) = '[a, b, c, d]

TupleToList (a, b, c, d, e) = '[a, b, c, d, e]

TupleToList (a, b, c, d, e, f) = '[a, b, c, d, e, f]

TupleToList (a, b, c, d, e, f, g) = '[a, b, c, d, e, f, g]

TupleToList (a, b, c, d, e, f, g, h) = '[a, b, c, d, e, f, g, h]

TupleToList (a, b, c, d, e, f, g, h, j) = '[a, b, c, d, e, f, g, h, j]

TupleToList (a, b, c, d, e, f, g, h, j, k) = '[a, b, c, d, e, f, g, h, j, k]

TupleToList (a, b, c, d, e, f, g, h, j, k, l) = '[a, b, c, d, e, f, g, h, j, k, l]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m) = '[a, b, c, d, e, f, g, h, j, k, l, m]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m, n) = '[a, b, c, d, e, f, g, h, j, k, l, m, n]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m, n, o) = '[a, b, c, d, e, f, g, h, j, k, l, m, n, o]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m, n, o, p) = '[a, b, c, d, e, f, g, h, j, k, l, m, n, o, p]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m, n, o, p, q) = '[a, b, c, d, e, f, g, h, j, k, l, m, n, o, p, q]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m, n, o, p, q, r) = '[a, b, c, d, e, f, g, h, j, k, l, m, n, o, p, q, r]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m, n, o, p, q, r, s) = '[a, b, c, d, e, f, g, h, j, k, l, m, n, o, p, q, r, s]

TupleToList a = '[a]