Data.Generics.Product.Internal.Types

Documentation

type family Children (ch :: Type) (a :: Type) :: [Type]

The children of a type are the types of its fields. The Children type family maps a type a to its set of children.

This type family is parameterized by a symbol ch (that can be declared as an empty data type). The symbol ChGeneric provides a default definition. You can create new symbols to override the set of children of abstract, non-generic types.

The following example declares a Custom symbol to redefine Children for some abstract types from the time library.

data Custom
type instance Children Custom a = ChildrenCustom a

type family ChildrenCustom (a :: Type) where
  ChildrenCustom DiffTime        = '[]
  ChildrenCustom NominalDiffTime = '[]
  -- Add more custom mappings here.

  ChildrenCustom a = Children ChGeneric a

To use this definition, replace types with typesUsing @Custom.

data ChGeneric

The default definition of Children. Primitive types from core libraries have no children, and other types are assumed to be Generic.

type family ChildrenDefault (a :: Type) :: [Type] where ...

Equations

ChildrenDefault Char = '[]

ChildrenDefault Double = '[]

ChildrenDefault Float = '[]

ChildrenDefault Integer = '[]

ChildrenDefault Int = '[]

ChildrenDefault Int8 = '[]

ChildrenDefault Int16 = '[]

ChildrenDefault Int32 = '[]

ChildrenDefault Int64 = '[]

ChildrenDefault Word = '[]

ChildrenDefault Word8 = '[]

ChildrenDefault Word16 = '[]

ChildrenDefault Word32 = '[]

ChildrenDefault Word64 = '[]

ChildrenDefault Text = '[]

ChildrenDefault (Param n _) = '[]

ChildrenDefault a = Defined (Rep a) (NoGeneric a '['Text "arising from a generic traversal.", 'Text "Either derive the instance, or define a custom traversal using HasTypesCustom"]) (ChildrenGeneric (Rep a) '[])

type family ChildrenGeneric (f :: k -> Type) (cs :: [Type]) :: [Type] where ...

Equations

ChildrenGeneric (M1 _ _ f) cs = ChildrenGeneric f cs

ChildrenGeneric (l :*: r) cs = ChildrenGeneric l (ChildrenGeneric r cs)

ChildrenGeneric (l :+: r) cs = ChildrenGeneric l (ChildrenGeneric r cs)

ChildrenGeneric (Rec0 a) cs = a ': cs

ChildrenGeneric _ cs = cs

type Interesting (ch :: Type) (a :: Type) (t :: Type) = Defined_list (Children ch t) (NoChildren ch t) (IsNothing (Interesting' ch a '[t] (Children ch t)))

type family NoChildren (ch :: Type) (a :: Type) :: Constraint where ...

Equations

NoChildren ch a = PrettyError '['Text "No type family instance for " ':<>: QuoteType (Children ch a), 'Text "arising from a traversal over " ':<>: QuoteType a, 'Text "with custom strategy " ':<>: QuoteType ch]

type family Interesting' (ch :: Type) (a :: Type) (seen :: [Type]) (ts :: [Type]) :: Maybe [Type] where ...

Equations

Interesting' ch _ seen '[] = 'Just seen

Interesting' ch a seen (t ': ts) = InterestingOr ch a (InterestingUnless ch a seen t (Elem t seen)) ts

type family InterestingUnless (ch :: Type) (a :: Type) (seen :: [Type]) (t :: Type) (alreadySeen :: Bool) :: Maybe [Type] where ...

Equations

InterestingUnless ch a seen a _ = 'Nothing

InterestingUnless ch a seen t 'True = 'Just seen

InterestingUnless ch a seen t 'False = Defined_list (Children ch t) (NoChildren ch t) (Interesting' ch a (t ': seen) (Children ch t))

type family InterestingOr (ch :: Type) (a :: Type) (seen' :: Maybe [Type]) (ts :: [Type]) :: Maybe [Type] where ...

Equations

InterestingOr ch a 'Nothing _ = 'Nothing

InterestingOr ch a ('Just seen) ts = Interesting' ch a seen ts

type family Elem a as where ...

Equations

Elem a (a ': _) = 'True

Elem a (_ ': as) = Elem a as

Elem a '[] = 'False

type family IsNothing a where ...

Equations

IsNothing ('Just _) = 'False

IsNothing 'Nothing = 'True

class HasTypes s a where

Minimal complete definition

Nothing

Methods

types_ :: Traversal' s a

data Void

class HasTypesUsing (ch :: Type) s t a b where

Since: 1.2.0.0

Methods

typesUsing_ :: Traversal s t a b

class HasTypesCustom (ch :: Type) s t a b where

By adding instances to this class, we can override the default behaviour in an ad-hoc manner. For example:

instance HasTypesCustom Custom Opaque Opaque String String where
  typesCustom f (Opaque str) = Opaque $ f str

Since: 1.2.0.0

Methods

typesCustom :: Traversal s t a b

This function should never be used directly, only to override the default traversal behaviour. To actually use the custom traversal strategy, see typesUsing. This is because typesUsing does additional optimisations, like ensuring that nodes with no relevant members will not be traversed at runtime.

class HasTypesOpt (ch :: Type) (p :: Bool) s t a b where

Methods

typesOpt :: Traversal s t a b

class GHasTypes ch s t a b where

Methods

gtypes_ :: Traversal (s x) (t x) a b