Data.Generics.Sum.Internal.Constructors

Description

Derive constructor-name-based prisms generically.

Documentation

class GAsConstructor (ctor :: Symbol) s t a b | ctor s -> a, ctor t -> b where

As AsConstructor but over generic representations as defined by GHC.Generics.

Methods

_GCtor :: Prism (s x) (t x) a b

type GAsConstructor' ctor s a = GAsConstructor ctor s s a a

type Context' ctor s a = (Context0 ctor s s a a, ErrorUnless ctor s (HasCtorP ctor (Rep s)))

class Context (ctor :: Symbol) s t a b | ctor s -> a, ctor t -> b

class Context_ (ctor :: Symbol) s t a b

type Context0 ctor s t a b = (Generic s, Generic t, GAsConstructor ctor (Rep s) (Rep t) a b, Defined (Rep s) (NoGeneric s '[(('Text "arising from a generic prism focusing on the " ':<>: QuoteType ctor) ':<>: 'Text " constructor of type ") ':<>: QuoteType a, 'Text "in " ':<>: QuoteType s]) (() :: Constraint))

derived0 :: forall ctor s t a b. Context0 ctor s t a b => Prism s t a b