Data.Generics.Sum.Internal.Typed

Description

Derive constructor-field-type-based prisms generically.

Documentation

type Context a s = (Generic s, ErrorUnlessOne a s (CollectPartialType (TupleToList a) (Rep s)), GAsType (Rep s) a, Defined (Rep s) (NoGeneric s '['Text "arising from a generic prism focusing on a constructor of type " ':<>: QuoteType a]) (() :: Constraint))

derived :: Context a s => Prism' s a

class GAsType (f :: Type -> Type) (as :: Type) where

As AsType but over generic representations as defined by GHC.Generics.

Methods

_GTyped :: Prism (f x) (f x) as as